contact router updated based on validation middleware

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require('../models/contactModel');
 
 //@desc - get all contacts
 //@route - GET - api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc - get a unique contact
 //@route - GET - api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -23,7 +23,7 @@ const getContact = asyncHandler(async (req, res)=>{
 
 //@desc - create a contact
 //@route - POST - api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res)=>{
     const {name, email, phone} = req.body;
     if(!name){
@@ -41,32 +41,44 @@ const createContact = asyncHandler(async (req, res)=>{
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
     res.status(201).json(contact);
 });
 
 //@desc - delete a contact
 //@route - DELETE - api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission for this operation")
+    }
+
     await Contact.findByIdAndDelete(req.params.id);
     res.status(204).json(contact);
 });
 
 //@desc - update a contact
 //@route - PUT - api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission for this operation")
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
